backend/requirements.txt
django>=5.0
django-cors-headers
PyPDF2
python-docx
